Wakeing up the MQTT

+++ misc/bridge_wsl_ports.ps1
@@ -1,3 +1,6 @@
+# Use like this: 
+# powershell.exe -File "[PATH_TO_SCRIPT]\brigde_wsl_ports.ps1"
+
 $ports = @(1883);
 
 # Get the WSL IP address using bash command
@@ -13,6 +16,7 @@ else {
   exit
 }
 
+
 # Set listen address
 $listenAddress = '0.0.0.0';
 
@@ -21,29 +25,36 @@ foreach ($port in $ports) {
   try {
     Invoke-Expression "netsh interface portproxy delete v4tov4 listenport=$port listenaddress=$listenAddress";
     Invoke-Expression "netsh interface portproxy add v4tov4 listenport=$port listenaddress=$listenAddress connectport=$port connectaddress=$wslAddress";
+    
+    Write-Host "Port forwarding set up for port $port" -ForegroundColor Green;
   }
   catch {
     Write-Host "Error while setting port forwarding for port $port" -ForegroundColor Red
-    exit
   }
 }
+
 
 # Set firewall rule name
 $fireWallDisplayName = 'WSL Port Forwarding';
 $portsStr = $ports -join ",";
 
-# Remove any existing firewall rules with the same display name
-try { 
-  Remove-NetFirewallRule -DisplayName $fireWallDisplayName 
+# Remove any existing firewall rules with the same display name if they exist
+# they should come from earlier iterations of this script
+if (Get-NetFirewallRule -DisplayName $fireWallDisplayName -ErrorAction SilentlyContinue) {
+  try {
+    Remove-NetFirewallRule -DisplayName $fireWallDisplayName
+  }
+  catch {
+    Write-Host "Error removing existing firewall rules" -ForegroundColor Red
+  }
 }
-catch {
-  Write-Host "Error removing existing firewall rules" -ForegroundColor Red
-}
+
 
 # Create outbound and inbound firewall rules
 try {
   New-NetFirewallRule -DisplayName $fireWallDisplayName -Direction Outbound -LocalPort $portsStr -Action Allow -Protocol TCP
   New-NetFirewallRule -DisplayName $fireWallDisplayName -Direction Inbound -LocalPort $portsStr -Action Allow -Protocol TCP
+  Write-Host "Firewall rules created for ports $portsStr" -ForegroundColor Green
 }
 catch {
   Write-Host "Error creating firewall rules" -ForegroundColor Red
